#7519 - SampleListService
