+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sample/services/AdditionalTestService.java
@@ -1,4 +1,4 @@
-package de.symeda.sormas.backend.sample;
+package de.symeda.sormas.backend.sample.services;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -28,6 +28,8 @@ import de.symeda.sormas.backend.common.CoreAdo;
 import de.symeda.sormas.backend.common.CriteriaBuilderHelper;
 import de.symeda.sormas.backend.contact.Contact;
 import de.symeda.sormas.backend.event.EventParticipant;
+import de.symeda.sormas.backend.sample.AdditionalTest;
+import de.symeda.sormas.backend.sample.Sample;
 import de.symeda.sormas.backend.user.User;
 import de.symeda.sormas.backend.util.QueryHelper;
