+++ sormas-backend/src/test/java/de/symeda/sormas/backend/labmessage/LabMessageFacadeEjbMappingTest.java
@@ -24,7 +24,7 @@ import de.symeda.sormas.api.sample.SampleMaterial;
 import de.symeda.sormas.api.sample.SampleReferenceDto;
 import de.symeda.sormas.api.sample.SpecimenCondition;
 import de.symeda.sormas.backend.sample.Sample;
-import de.symeda.sormas.backend.sample.SampleService;
+import de.symeda.sormas.backend.sample.services.SampleService;
 import junit.framework.TestCase;
 
 @RunWith(MockitoJUnitRunner.class)
